Unfinished: analyse_data_ratios and _INTcontrol
The actual analysis for left/right sides is finished, I'm just trying to make collapse across coherences a thing. If this script doesn't work, delete things which start with "if cohs_collapsed == 1" because those are the unfinished bits.

Also, the second script (which is highly similar to the first one) controls for integration length. If the first script doesn't work, you could always slightly change the latter back into the former (as an alternative to the above deletion of IF statements).